refactor(): moving validator to correct package
