+++ utility/synapse-utilities-common/src/main/java/io/americanexpress/synapse/utilities/common/email/EmailAddressUtils.java
@@ -1,12 +1,25 @@
+/*
+ * Copyright 2020 American Express Travel Related Services Company, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package io.americanexpress.synapse.utilities.common.email;
 
 import org.apache.commons.lang3.StringUtils;
 import java.util.regex.Pattern;
 
 /**
- * {@code EmailAddressValidator} is used for email address validation.
+ * {@code EmailAddressUtils} is used for email address validation.
  */
-public class EmailAddressValidator {
+public class EmailAddressUtils {
 
     /**
      * The email address regex.
@@ -21,7 +34,7 @@ public class EmailAddressValidator {
     /**
      * Private constructor to prevent instantiation of utility class.
      */
-    private EmailAddressValidator() {}
+    private EmailAddressUtils() {}
 
     /**
      * Checks if an email address is valid.
