+++ utility/synapse-utilities-common/src/main/java/io/americanexpress/synapse/utilities/common/validator/EmailAddress.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2020 American Express Travel Related Services Company, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package io.americanexpress.synapse.utilities.common.validator;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * {@code EmailAddress} annotation used to validate email address field in models.
+ */
+@Documented
+@Constraint(validatedBy = EmailAddressValidator.class)
+@Target( { ElementType.FIELD })
+@Retention(RetentionPolicy.RUNTIME)
+public @interface EmailAddress {
+
+    String message() default "Invalid email address";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}
